+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - PD/ED characteristics with resistant modifier are now actually resistant.  Previously the resistant modifier was ignored.  Also PD/ED purchased as characteristics now show in DEFENSES tab for easy reference. [#1063](https://github.com/dmdorman/hero6e-foundryvtt/issues/1063)
 - Fix where the CSL selections were duplciated during AOE template placement. [#1064](https://github.com/dmdorman/hero6e-foundryvtt/issues/1064)
+- Fix where DCV temporary bonuses from CSLs and other active effects were expiring on our segment instead of on our phase. [#1061](https://github.com/dmdorman/hero6e-foundryvtt/issues/1061)
 
 
 ## Version 3.0.76 [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)